feat: added operations with native C types

+++ include/utils.h
@@ -48,7 +48,7 @@ do {\
 #define strtol_or_error(__str, __result, __on_error) \
 do {\
     char* __endptr; \
-    int32_t __value = strtol(__str, &__endptr, 10); \
+    long __value = strtol(__str, &__endptr, 10); \
     if ((errno == ERANGE && (__value == LONG_MAX || __value == LONG_MIN)) || __str == __endptr || *__endptr != '\0') {\
         __on_error \
     } \
@@ -60,7 +60,7 @@ do {\
 #define strtoll_or_error(__str, __result, __on_error) \
 do {\
     char* __endptr; \
-    int64_t __value = strtoll(__str, &__endptr, 10); \
+    long long __value = strtoll(__str, &__endptr, 10); \
     if ((errno == ERANGE && (__value == LLONG_MAX || __value == LLONG_MIN)) || __str == __endptr || *__endptr != '\0') {\
         __on_error \
     } \
@@ -72,7 +72,7 @@ do {\
 #define strtoull_or_error(__str, __result, __on_error) \
 do {\
     char* __endptr; \
-    u_int64_t __value = strtoull(__str, &__endptr, 10); \
+    unsigned long long __value = strtoull(__str, &__endptr, 10); \
     if ((errno == ERANGE && (__value == ULLONG_MAX)) || __str == __endptr || *__endptr != '\0') {\
         __on_error \
     } \
@@ -84,7 +84,7 @@ do {\
 #define strtoul_or_error(__str, __result, __on_error) \
 do {\
     char* __endptr; \
-    u_int32_t __value = strtoul(__str, &__endptr, 10); \
+    unsigned long __value = strtoul(__str, &__endptr, 10); \
     if ((errno == ERANGE && (__value == ULONG_MAX)) || __str == __endptr || *__endptr != '\0') {\
         __on_error \
     } \
